Revert change of deliverData() iLength argument from type 'int' to type 'size_t'. This change was preventing the notification of stream read completion, as indicated by a read length of '-1', from being detected. (XECS-1405)
git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@12633 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXmediaLib/include/mp/StreamHttpDataSource.h
@@ -102,7 +102,7 @@ public:
      //!param iLocation - The desired seek location
 
 
-   UtlBoolean deliverData(char *szData, size_t iLength, size_t iMaxLength) ;
+   UtlBoolean deliverData(char *szData, int iLength, size_t iMaxLength) ;
      //:Callback routine that is invoked whenever new data is available from
      //:http socket.
